add delay option

+++ src/bootstrap/opts.rs
@@ -12,8 +12,8 @@ pub struct Opts {
     #[clap(long, default_value = "25565")]
     pub port: u16,
 
-    #[clap(long, default_value = "0")]
-    pub delay: u16,
+    #[clap(short, long, default_value = "500")]
+    pub delay: u64,
 
     #[clap(short, long)]
     pub online: bool,
